requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cloudinary==1.40.0
crispy-bootstrap4==2024.1
cssselect2==0.7.0
Django==5.0.6
django-admin-confirm==1.0.0
django-admin-interface==0.28.7
django-axes==6.5.0
django-cloudinary-storage==0.3.0
django-colorfield==0.11.0
django-crispy-forms==2.2
django-environ==0.11.2
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
fonttools==4.53.0
html5lib==1.1
idna==3.7
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
python-slugify==8.0.4
reportlab==4.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tinycss2==1.3.0
tzdata==2024.1
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.7.0
zopfli==0.2.3
